Add progress test to filesystem_simulator
Summary:
In filesystem_simulator, in addition to checking
consistency, add a check to ensure that committed
transactions have been correctly persisted,
for example, a file that is created and never deleted
should exist, and a file that is deleted should not.
This helps verify that the database does not lose
committed data.

Test Plan:
Make the updated test pass with a local single node
cockroach cluster.

Manually do a negative test by deleting a file
from behind the test and ensuring the test fails.

Reviewers: nishant.gupta, jj, garvit

Reviewed By: garvit

Tags: #cly

Differential Revision: https://phabricator.rubrik.com/D27748